+++ utils/imageio.c
@@ -973,12 +973,12 @@ static IMAGE *TiffReadImage(const char *fname, int frame0)
           bitmap = (unsigned char *)calloc(scanlinesize, sizeof(unsigned char));
 
           memmove(bitmap, buf, scanlinesize);
-          for (col = b = 0; b < scanlinesize; b++, col += 8) {
+          for (col = b = 0; b < (unsigned int)scanlinesize; b++, col += 8) {
             byte_ = bitmap[b];
             if (byte_ > 0) DiagBreak();
             if (fillorder == FILLORDER_LSB2MSB) {
               for (bitmask = 0x01, bit = 0; bit < 8; bit++) {
-                if (col + bit == Gx && index == Gy) DiagBreak();
+                if (col + bit == (unsigned int)Gx && index == Gy) DiagBreak();
                 *IMAGEpix(I, col + bit, index) = ((byte_ & bitmask) > 0);
                 bitmask = bitmask << 1;
               }
@@ -986,7 +986,7 @@ static IMAGE *TiffReadImage(const char *fname, int frame0)
             else  // fillorder == FILLORDER_MSB2LSB
             {
               for (bitmask = 0x01 << 7, bit = 0; bit < 8; bit++) {
-                if (col + bit == Gx && index == Gy) DiagBreak();
+                if (col + bit == (unsigned int)Gx && index == Gy) DiagBreak();
                 *IMAGEpix(I, col + bit, index) = ((byte_ & bitmask) > 0);
                 bitmask = bitmask >> 1;
               }
